cleaned up preload
