+++ client/preload_assets/audio.js
@@ -1,0 +1,5 @@
+/* global D6Dungeon */
+
+export const preloadAudio = () => {
+  D6Dungeon.game.load.audio('boss-battle', 'assets/audio/boss-battle-2.mp3');
+};
