+++ client/preload_assets/index.js
@@ -1,0 +1,5 @@
+import { preloadAudio } from './audio';
+import { preloadSprites } from './sprites';
+import { preloadTilemaps } from './tilemaps';
+
+export { preloadAudio, preloadSprites, preloadTilemaps };
