+++ public/assets/character_sprites/red-horned-bee.json
@@ -1,0 +1,28 @@
+{"frames": {
+
+"sprite1.png":
+{
+	"frame": {"x":1,"y":1,"w":75,"h":63},
+	"rotated": false,
+	"trimmed": true,
+	"spriteSourceSize": {"x":2,"y":3,"w":75,"h":63},
+	"sourceSize": {"w":81,"h":68}
+},
+"sprite2.png":
+{
+	"frame": {"x":1,"y":66,"w":75,"h":58},
+	"rotated": false,
+	"trimmed": true,
+	"spriteSourceSize": {"x":2,"y":8,"w":75,"h":58},
+	"sourceSize": {"w":81,"h":68}
+}},
+"meta": {
+	"app": "https://www.codeandweb.com/texturepacker",
+	"version": "1.0",
+	"image": "red-horned-bee.png",
+	"format": "RGBA8888",
+	"size": {"w":77,"h":125},
+	"scale": "0.125",
+	"smartupdate": "$TexturePacker:SmartUpdate:fc559bceecaee93fdaa21c0a89f6fd6f:4e5e5af652d40f74e544631468c8789a:05d8d419f707dbc1c049a443c1b9b220$"
+}
+}
